Move utility methods into separated package.
